API/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
azure-cognitiveservices-speech==1.38.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
diff-match-patch==20230430
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
django-import-export==4.1.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
elevenlabs==1.6.1
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.1
openai==1.37.1
openpyxl==3.1.5
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tablib==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==12.0
